Separate array functions into a file
