+++ src/utils/common.js
@@ -21,35 +21,12 @@ const formatDate = (date) => DateFormatter.format(date);
 
 const isFirst = (index) => index === 0;
 
-const getFirst = (array) => array[0];
-
-const getLast = (array) => array[array.length - 1];
-
 const getDate = (date) => new Intl.DateTimeFormat(`en-US`).format(date);
-
-const getUnique = (array) => [...new Set(array)];
-
-// Sorting
-
-const sortByDurationInDescendingOrder = (a, b) => (b.endTime - b.startTime) - (a.endTime - a.startTime);
-
-const sortByPriceInDescendingOrder = (a, b) => b.price - a.price;
-
-const sortByStartTimeInAscendingOrder = (a, b) => a.startTime - b.startTime;
-
-const sortPurely = (collection, iterate) => collection.slice().sort(iterate);
 
 export {
   getTwoDigitFormat,
   formatTime,
   formatDate,
   isFirst,
-  getFirst,
-  getLast,
   getDate,
-  getUnique,
-  sortByDurationInDescendingOrder,
-  sortByPriceInDescendingOrder,
-  sortByStartTimeInAscendingOrder,
-  sortPurely
 };
